# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample HTTP package with Swagger annotations.",
        "title": "VK-Filmoteka",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/http/v1/actor": {
            "get": {
                "description": "Можно задавать разные query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Получение актера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "текущая страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "кол-во актеров на странице",
                        "name": "page-count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Вернуть вместе со связями (film)",
                        "name": "connection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "получаемые актеры",
                        "schema": {
                            "$ref": "#/definitions/appDto.ActorGetByQueryResult"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Обновление актера [Админы]",
                "parameters": [
                    {
                        "description": "Данные актера",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные актера",
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Создание актера [Админы]",
                "parameters": [
                    {
                        "description": "Данные актера",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appDto.CreateActorUseCaseDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные созданного актера",
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Доступно только админам, ничего ответом не возвращает",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Удаление актера [Админы]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id удаляемого пользователья",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/http/v1/film": {
            "get": {
                "description": "Можно задавать разные query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Получение фильма",
                "parameters": [
                    {
                        "type": "string",
                        "description": "текущая страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "кол-во актеров на странице",
                        "name": "page-count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Вернуть вместе со связями (film)",
                        "name": "connection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc либо desc",
                        "name": "order-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "поле по которому сортируют (rate, name, release_date)",
                        "name": "order-field",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "получаемые фильмы",
                        "schema": {
                            "$ref": "#/definitions/appDto.FilmGetByQueryResult"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Создание фильма [Админы]",
                "parameters": [
                    {
                        "description": "Данные фильма",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные обновленного фильма",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Создание фильма [Админы]",
                "parameters": [
                    {
                        "description": "Данные фильма",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appDto.CreateFilmUseCaseDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные созданного фильма",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Доступно только админам, ничего не возвоащает",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Создание фильма [Админы]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id удаляемого фильма",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
        "/http/v1/film/search": {
            "get": {
                "description": "поиск по имени и имени актеров",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Поиск фильма",
                "parameters": [
                    {
                        "type": "string",
                        "description": "параметр поиска",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "получаемые фильмы",
                        "schema": {
                            "$ref": "#/definitions/appDto.FilmGetByQueryResult"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "aggregate.ActorAggregate": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/model.Actor"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Film"
                    }
                }
            }
        },
        "aggregate.FilmAggregate": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Actor"
                    }
                },
                "film": {
                    "$ref": "#/definitions/model.Film"
                }
            }
        },
        "appDto.ActorGetByQueryResult": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/aggregate.ActorAggregate"
                    }
                },
                "pageCount": {
                    "type": "integer"
                }
            }
        },
        "appDto.CreateActorUseCaseDto": {
            "type": "object",
            "required": [
                "birthday",
                "gender",
                "name"
            ],
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "appDto.CreateFilmUseCaseDto": {
            "type": "object",
            "required": [
                "name",
                "release"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "rate": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release": {
                    "type": "string"
                }
            }
        },
        "appDto.FilmGetByQueryResult": {
            "type": "object",
            "properties": {
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/aggregate.FilmAggregate"
                    }
                },
                "pageCount": {
                    "type": "integer"
                }
            }
        },
...
        "appErrors.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Actor": {
            "type": "object",
            "required": [
                "birhday",
                "gender",
                "id",
                "name"
            ],
            "properties": {
                "birhday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "model.Film": {
            "type": "object",
            "required": [
                "id",
                "name",
                "release"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "rate": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release": {
                    "type": "string"
                }
            }
        }
    }
}